Revert "remove unused gethitssize, revert getmemsize to size_t"
This reverts commit c95c1050e14e2ee3396d5165bee6537b0a916b50.

+++ src/Simulation/MolflowSimFacet.cpp
@@ -238,7 +238,6 @@ bool MolflowSimFacet::InitializeLinkAndVolatile(const int  id)
 * \param nbMoments amount of moments (TD simulations are currently not used for Molflow)
 * \return calculated size of the facet hits
 */
-/*
 int MolflowSimFacet::GetHitsSize(int nbMoments) const {
     // Init with base size
     int hits_size = SimulationFacet::GetHitsSize(nbMoments);
@@ -250,10 +249,9 @@ int MolflowSimFacet::GetHitsSize(int nbMoments) const {
 size_t MolflowSimFacet::GetMemSize() const {
 
     // Init with base size
-    size_t mem_size = SimulationFacet::GetMemSize();
+    int mem_size = SimulationFacet::GetMemSize();
 
     mem_size += sizeof (double) * ogMap.outgassingMap.capacity();
     mem_size += angleMap.GetMemSize();
     return mem_size;
 }
-*/
